Handle a missing lib/ directory more seamlessly.

+++ nose.sh
@@ -25,6 +25,11 @@ done
 
 # We use without-sandbox, so that we properly load all the modules/tests
 cat app.yaml | sed 's/runtime: vm/runtime: python27/' > app-nose.yaml
-rm -rf lib/tests # this is pulled in via twilio, and messes with our excludes
-MODULES=$(find lib -maxdepth 1 | grep -v info | cut -f2- -d/ | sed 's/\.py//' | paste -s -d "|" -)
+# If our lib/ directory exists, manage it properly, otherwise assume the modules are all installed globally
+if [ -d lib ]; then
+	rm -rf lib/tests # this is pulled in via twilio, and messes with our excludes
+	MODULES=$(find lib -maxdepth 1 | grep -v info | cut -f2- -d/ | sed 's/\.py//' | paste -s -d "|" -)
+else
+	MODULES=""
+fi
 PYTHONPATH=lib $COVERAGE_PREFIX `which nosetests` --with-gae --gae-application=app-nose.yaml --exclude="$MODULES" ${NOSE_ARGS[@]}
